backend/requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
pydantic-settings
pytest
httpx
python-multipart
python-dotenv
